refactor : 패키지 이름을 모두 소문자로 변경
