--- src/main/java/racingGame/GameApplication.java
@@ -1,12 +1,0 @@
-package racingGame;
-
-import racingGame.business.Process;
-
-public class GameApplication {
-    
-    public static void main(String[] args) {
-        Process process = new Process();
-        ResultView.printAllResult(process.run(InputView.inputCarName(), InputView.inputMoveNum()));
-    }
-
-}
