--- src/main/java/racingGame/model/GameResult.java
@@ -1,16 +1,0 @@
-package racingGame.model;
-
-
-import java.util.ArrayList;
-import java.util.List;
-
-public record GameResult(List<ProgressRecord> progressRecords) {
-    
-    public GameResult(List<ProgressRecord> progressRecords) {
-        this.progressRecords = new ArrayList<>(progressRecords);
-    }
-    
-    public JoinCars findGameWinners() {
-        return progressRecords.getLast().joinCars().findWinners();
-    }
-}
