--- src/main/java/racingGame/response/GameResponse.java
@@ -1,8 +1,0 @@
-package racingGame.response;
-
-import racingGame.model.GameResult;
-import racingGame.model.JoinCars;
-
-public record GameResponse(GameResult gameResult, JoinCars winners) {
-
-}
