+++ src/main/java/racinggame/GameApplication.java
@@ -1,0 +1,12 @@
+package racinggame;
+
+import racinggame.business.Process;
+
+public class GameApplication {
+    
+    public static void main(String[] args) {
+        Process process = new Process();
+        ResultView.printAllResult(process.run(InputView.inputCarName(), InputView.inputMoveNum()));
+    }
+
+}
