+++ src/main/java/racinggame/business/Process.java
@@ -1,0 +1,14 @@
+package racinggame.business;
+
+import racinggame.model.*;
+import racinggame.response.GameResponse;
+
+public class Process {
+    
+    public GameResponse run(String[] names, int moves) {
+        JoinCars joinCars = CarFactory.createCars(new NonNegativeReadyCars(names));
+        GameResult gameResult = new GameMove(new NonNegativeMoves(moves)).moveCar(joinCars);
+        JoinCars gameWinners = gameResult.findGameWinners();
+        return new GameResponse(gameResult, gameWinners);
+    }
+}
