+++ src/main/java/racinggame/model/GameResult.java
@@ -1,0 +1,16 @@
+package racinggame.model;
+
+
+import java.util.ArrayList;
+import java.util.List;
+
+public record GameResult(List<ProgressRecord> progressRecords) {
+    
+    public GameResult(List<ProgressRecord> progressRecords) {
+        this.progressRecords = new ArrayList<>(progressRecords);
+    }
+    
+    public JoinCars findGameWinners() {
+        return progressRecords.getLast().joinCars().findWinners();
+    }
+}
