+++ src/main/java/racinggame/response/GameResponse.java
@@ -1,0 +1,8 @@
+package racinggame.response;
+
+import racinggame.model.GameResult;
+import racinggame.model.JoinCars;
+
+public record GameResponse(GameResult gameResult, JoinCars winners) {
+
+}
