+++ src/test/java/racinggame/model/CarTest.java
@@ -1,0 +1,81 @@
+package racinggame.model;
+
+import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
+
+class CarTest {
+    
+    Car car;
+    
+    @BeforeEach
+    void setUp() {
+        String name = "mo";
+        car = new Car(name);
+    }
+    
+    @Test
+    void carGenerateTest() {
+        assertThat(car);
+    }
+
+    @Test
+    void carsNameBlankValidateTest() {
+        assertThatThrownBy(() -> {
+            new Car("");
+        }).isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("%s", "자동차 이름을 입력하거나, 5글자 이하이름을 쓰시오");
+    }
+
+    @Test
+    void carsNameOverLetterValidateTest() {
+        assertThatThrownBy(() -> {
+            new Car("123456");
+        }).isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("%s", "자동차 이름을 입력하거나, 5글자 이하이름을 쓰시오");
+    }
+
+    @Test
+    void carShowNameTest() {
+        assertThat(car.getName()).isEqualTo("mo");
+    }
+
+    @Test
+    void carForwardTest() {
+        int num = 5;
+        car.forward(num);
+        car.forward(num);
+        assertThat(car.forward(num)).isEqualTo(3);
+    }
+    
+    @ParameterizedTest
+    @ValueSource(ints = {4, 8})
+    void carIsForward(int num) {
+        assertThat(car.forward(num)).isEqualTo(1);
+    }
+    
+    @ParameterizedTest
+    @ValueSource(ints = {0, 1, 3})
+    void carIsNotForward(int num) {
+        assertThat(car.forward(num)).isEqualTo(0);
+    }
+    
+    @Test
+    void compareLocationTest() {
+        String name = "mo";
+        car = new Car(name,6);
+        assertThat(car.compareLocation(5)).isEqualTo(6);
+    }
+    
+    @Test
+    void isSameLocationTest() {
+        String name = "mo";
+        car = new Car(name,5);
+        assertThat(car.isSameLocation(5)).isTrue();
+    }
+    
+}
